refactor(execution): move CyclesAccountManagerConfig SEV tests to cycles_account_manager
The two tests asserting reference_subnet_size for SEV enabled/disabled are
a better fit alongside the existing reference_subnet_size tests in
cycles_account_manager. SubnetConfig-level tests remain in subnet_config.rs.
